Create universal function to load personalized ads - Closes #68
